Idea for wiki: use ikiwiki or something similar.
I really like the concept of a wiki compiler and of a wiki backed by a
distributed revision control system that allows you to edit the pages
offline in your favorite editor and merge edits using the full power
of whatever merge tool you prefer.

Unfortunately ikiwiki is written in Perl (of which I don't understand
much) and git is the preferred back end. So I guess we have to come up
with our own system for this.