Implemented a multi-broadcast method.
The multi-broadcast method replaces the single-broadcast method. A
single call allows every player to broadcast at most one message by
taking a list of senders instead of a single sender as an argument. To
do a normal broadcast as before, simply place the single sender as the
only element in the list of senders and everything else works as it
did before.