Use pack format string for signed input.
The hash function returns a Python integer, which is a signed 32 bit
number. In Python 2.4, struct.pack("L", x) would automatically wrap a
negative number x to 2**32 - x, but in Python 2.5 this triggers a
DeprecationWarning and gives four \x00 bytes as the result. Using the
"l" format solves the problem.